\name{candlesticks-package}
\alias{candlesticks-package}
\alias{candlesticks}
\docType{package}
\title{
Functions for Candlestick Pattern Recognition
}
\description{
This package contains functions to detect the most popular japanese candlestick formations like Doji, Engulfing Pattern, Harami, etc. out of OHLC price data.
}
\details{
\tabular{ll}{
Package: \tab candlesticks\cr
Type: \tab Package\cr
Version: \tab 0.1-1\cr
Date: \tab 2012-01-10\cr
Depends: \tab xts,quantmod,TTR\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Andreas Voellenklee

Maintainer: Andreas Voellenklee <wotuzu17@gmail.com>
}
\references{
  The following sites were used to code/document this package:\cr
  \url{http://www.onlinetradingconcepts.com/TechnicalAnalysis/Candlesticks/CandlestickBasics.html}\cr
  \url{http://candlesticker.com/}\cr
  \url{http://thepatternsite.com/CandleEntry.html}\cr
}
\keyword{ package }
\seealso{

}
\examples{

}
